+++ templates/default.lua
@@ -1,8 +1,7 @@
 local template = {
-    collection = "Templating Engine",
     name = "Default",
     category = "menu.templates",
-    settings = { target = "templates", enabled = false },
+    settings = { target = "templates", enabled = true },
 
     menu = {
         aggregate = {